✨ Quit on Ctrl+C
For a terminal-based app, some users expect Ctrl+C to work as universal
"quit this". Let's implement this. Note that we only check for Control
to be *in* the modifier set (but not exclusively so), so Alt+Control+C
or similar would also quit. That can be fixed separately if there's ever
a need for that.

Fixes #39