💥 Color scheme specification now mandatory with git hook
When running as the git commit hook, we now require the color scheme to
be explicitly specified. If it's not specified, we ask the user to
re-install the hook (which is just one command). Otherwise, the
autodection will kick in and on some operating systems, we face the
garbage input issue (#69).